add support for Pimoroni pHAT
fixes #14
