+++ report.py
@@ -9,59 +9,9 @@ import logging
 import sys
 import time
 
-try:
-    import board
-    import busio
-    import digitalio
-except NotImplementedError as exc:
-    print(f"Will only support running with -o: {exc}")
-from adafruit_epd.ssd1680 import Adafruit_SSD1680
-
-from display import Display
+from display import get_e_ink_display
 from logutil import LogLevelAction
-
-
-def get_e_ink_display(height, width):
-    """
-    :return: eInk display instance
-    """
-    # Create the SPI device and pins we will need.
-    spi = busio.SPI(board.SCK, MOSI=board.MOSI, MISO=board.MISO)
-    ecs = digitalio.DigitalInOut(board.CE0)
-    # pylint: disable=invalid-name
-    dc = digitalio.DigitalInOut(board.D22)
-    rst = digitalio.DigitalInOut(board.D27)
-    busy = digitalio.DigitalInOut(board.D17)
-
-    display = Adafruit_SSD1680(
-        height,
-        width,
-        spi,
-        cs_pin=ecs,
-        dc_pin=dc,
-        sramcs_pin=None,
-        rst_pin=rst,
-        busy_pin=busy,
-    )
-
-    display.rotation = 1
-
-    return display
-
-
-def update_e_ink_display(display, image):
-    """
-    Display image.
-    :param display:
-    :param image:
-    :return:
-    """
-    logger = logging.getLogger(__name__)
-
-    logger.debug("display in progress")
-    display.image(image)
-    display.display()
-    logger.debug("display done")
+from metrics_drawer import MetricsDrawer
 
 
 def parse_args():
@@ -133,7 +83,7 @@ def main():
     display_height = 122
     display_width = 250
 
-    display = Display(
+    drawer = MetricsDrawer(
         args.url,
         display_width,
         display_height,
@@ -142,12 +92,15 @@ def main():
         args.temp_sensor_name,
     )
     if args.output:
-        image = display.draw_image()
+        image = drawer.draw_image()
         image.save(args.output)
         return
 
-    e_display = get_e_ink_display(display_height, display_width)
-    display = Display(
+    e_display = get_e_ink_display()
+    if e_display is None:
+        logger.error("No display detected")
+        sys.exit(1)
+    drawer = MetricsDrawer(
         args.url,
         e_display.width,
         e_display.height,
@@ -156,8 +109,8 @@ def main():
         args.temp_sensor_name,
     )
     while True:
-        image = display.draw_image()
-        update_e_ink_display(e_display, image)
+        image = drawer.draw_image()
+        e_display.update(image)
         time.sleep(args.timeout)
