+++ requirements.txt
@@ -5,3 +5,4 @@ pillow
 requests
 wheel
 prometheus-api-client
+inky
